rename ddal-demo to ddal-example
