--- ddal-demo/ddal-demo-demo0/src/main/resources/context.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:tx="http://www.springframework.org/schema/tx" xmlns:context="http://www.springframework.org/schema/context"
-       xsi:schemaLocation="
-          http://www.springframework.org/schema/beans
-          http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
-          http://www.springframework.org/schema/tx
-          http://www.springframework.org/schema/tx/spring-tx-3.1.xsd
-          http://www.springframework.org/schema/context
-          http://www.springframework.org/schema/context/spring-context-3.1.xsd"
-       default-autowire="byName">
-
-    <context:component-scan base-package="org.hellojavaer.ddal.demo.demo0"/>
-
-    <bean name="dataSource" class="org.hellojavaer.ddal.datasource.DefaultDDALDataSource">
-        <constructor-arg index="0" value="jdbc:ddal:thick:classpath:/datasource.xml"/>
-    </bean>
-
-    <bean id="txManager" class="org.springframework.jdbc.datasource.DataSourceTransactionManager">
-        <property name="defaultTimeout" value="6000"/>
-        <property name="dataSource" ref="dataSource"/>
-    </bean>
-    <tx:annotation-driven transaction-manager="txManager"/>
-
-    <bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
-        <property name="dataSource" ref="dataSource"/>
-        <property name="mapperLocations" value="classpath*:/mapper/user.xml"/>
-        <property name="configLocation" value="/mybatis.xml"/>
-    </bean>
-    <bean id="sqlSession" class="org.mybatis.spring.SqlSessionTemplate">
-        <constructor-arg index="0" ref="sqlSessionFactory"/>
-    </bean>
-
-</beans>
