--- ddal-demo/ddal-demo-demo1/src/main/java/org/hellojavaer/ddal/demo/demo1/dao/impl/RoleDaoImpl.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2017-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hellojavaer.ddal.demo.demo1.dao.impl;
-
-import org.hellojavaer.ddal.demo.demo1.dao.RoleDao;
-import org.hellojavaer.ddal.demo.demo1.entitry.RoleEntity;
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
-
-/**
- *
- * @author <a href="mailto:hellojavaer@gmail.com">Kaiming Zou</a>,created on 22/07/2017.
- */
-@Repository
-public class RoleDaoImpl implements RoleDao {
-
-    @Autowired
-    private SqlSessionTemplate sqlSessionTemplate;
-
-    @Override
-    public Long add(RoleEntity roleEntity) {
-        sqlSessionTemplate.insert("role.add", roleEntity);
-        return roleEntity.getId();
-    }
-
-    @Override
-    public int deleteById(Long id) {
-        return sqlSessionTemplate.delete("role.deleteById", id);
-    }
-
-    @Override
-    public int updateById(RoleEntity roleEntity) {
-        return sqlSessionTemplate.update("role.updateById", roleEntity);
-    }
-
-    @Override
-    public RoleEntity getById(Long id) {
-        return sqlSessionTemplate.selectOne("role.getById", id);
-    }
-}
