+++ ddal-example/ddal-example-example0/src/main/resources/log4j.xml
