+++ ddal-example/ddal-example-example1/src/main/resources/log4j.xml
