+++ pom.xml
@@ -22,7 +22,7 @@
         <module>ddal-datasource</module>
         <module>ddal-bom</module>
         <module>ddal-server</module>
-        <module>ddal-demo</module>
+        <module>ddal-example</module>
     </modules>
 
     <developers>
